requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
sqlmodel==0.0.16
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
telethon==1.34.0
cryptography==42.0.5
aiofiles==23.2.1
pydantic>=2.4.1,<2.6
pydantic-settings==2.1.0
openai==1.12.0
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
qrcode==7.4.2
